--- webapp/imports/api.old/documents/server/publications.js
@@ -1,4 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { Documents } from '../documents';
-
-Meteor.publish('documents', () => Documents.find());
