--- webapp/imports/api.old/posts/server/publications.js
@@ -1,8 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { Posts } from '../posts';
-
-//Meteor.publish('posts', () => Posts.find());
-
-Meteor.publish('posts', function(){
-  return Posts.find()
-});
