--- webapp/imports/api.old/statistics/server/publications.js
@@ -1,4 +1,0 @@
-import { Meteor } from 'meteor/meteor';
-import { Statistics } from '../statistics';
-
-Meteor.publish('statistics', () => Statistics.find());
