--- webapp/imports/api.old/topics/server/publications.js
@@ -1,3 +1,0 @@
-import { Topics } from '/imports/api/topics/topics';
-
-Meteor.publish('topics', () => Topics.find());
