+++ webapp/imports/ui/layouts/GlassLayout.jsx
@@ -56,7 +56,7 @@ export class GlassLayout extends React.Component {
       card: {
         title: 'Please log in',
         subtitle: 'Basic User',
-        avatar: '/noAvatar.png'
+        avatar: 'https://placeimg.com/80/80/animals'
       }
     };
