--- webapp/server/initialize.users.js
@@ -1,5 +1,0 @@
-
-// https://www.hl7.org/fhir/riskassessment-example-prognosis.json.html
-// https://www.hl7.org/fhir/riskassessment-example-cardiac.json.html
-import { Accounts } from 'meteor/accounts-base';
-import { Meteor } from 'meteor/meteor';
