+++ webapp/tests/nightwatch/walkthroughs/accounts.js
@@ -4,13 +4,6 @@
 
 module.exports = {
   tags: ['accounts', 'passwords', 'users', 'entry'],
-  before: function(client){
-    client
-      .url("http://localhost:3000").pause(3000)
-      .executeAsync(function(data){
-        Meteor.call('dropTestUsers');
-      });
-  },
   'User can sign up.': function (client) {
     client.resizeWindow(1200, 1024);
